Change doubles to floats

+++ mathlib/ErrorFunc.cpp
@@ -4,10 +4,10 @@
 
 using namespace mathlib;
 
-double ErrorFunc::categoricalCrossentropy(const Matrix& predictions, const Matrix& labels)
+float ErrorFunc::categoricalCrossentropy(const Matrix& predictions, const Matrix& labels)
 {
     // categorical crossentropy = - sum(labels * log(ouputs))
-    std::vector<double> multLabels;
+    std::vector<float> multLabels;
     for (int i = 0; i < labels.getRows(); i++)
     {   
         for (int j = 0; j < labels.getCols(); j++)
@@ -15,6 +15,6 @@ double ErrorFunc::categoricalCrossentropy(const Matrix& predictions, const Matri
             multLabels.emplace_back(labels(i, j) * std::log(predictions(i, j)));
         }
     }
-    return - std::accumulate(multLabels.begin(), multLabels.end(), 0.0);  
+    return - std::accumulate(multLabels.begin(), multLabels.end(), 0.f);  
 }
